Uploading Second C# Homework to GitHub
Uploading the second C# Homework about even and odd numbers. Basically checking if a number can be divided by 2 and the result in the substraction is zero (even) or if that is not the case and the number does not have any half (odd).

Second Homework is on the way.